requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-appconf==1.0.4
django-ckeditor==6.1.0
django-compressor==2.4.1
django-filter==21.1
django-js-asset==1.2.2
idna==3.2
Pillow==8.3.2
pytz==2021.1
rcssmin==1.0.6
requests==2.26.0
rjsmin==1.1.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
